Add user confirmation for rm command

+++ internal/cmd/rm.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"fmt"
 	"os"
+	"strings"
 )
 
 type RmCommand struct{}
@@ -21,10 +22,25 @@ func (c *RmCommand) Execute(args []string) (bool, error) {
 	}
 
 	path := args[0]
-	err := os.Remove(path)
-	if err != nil {
-		return false, fmt.Errorf("error removing the file: %s", err)
+	fmt.Printf("Are you sure you want to remove '%s'?\n", path)
+	fmt.Println("Y/N?")
+
+	var input string
+	fmt.Scan(&input)
+	input = strings.ToUpper(strings.TrimSpace(input))
+	switch input {
+	case "Y":
+		err := os.Remove(path)
+		if err != nil {
+			return true, fmt.Errorf("error removing the file: %s", err)
+		}
+		fmt.Printf(" - %s - file deleted successfully\n", path)
+		return true, nil
+	case "N":
+		return true, fmt.Errorf("you chose not to delete the file")
+	default:
+		fmt.Println("Please enter Y or N")
 	}
-	fmt.Printf(" - %s - file deleted successfully\n", path)
+
 	return false, nil
 }
